requirements.txt
aniso8601==3.0.2
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.1.0
openapi-codec==1.3.2
promise==2.2.1
public==1.0.3
PyJWT==1.6.4
PyMySQL==0.9.3
pytz==2018.5
requests==2.21.0
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
